refactor(core): extract core_graph boundary to break index/schema/intelligence/conventions cycle
